pyproject.toml
[tool.poetry]
name = "coachai-api"
version = "0.1.0"
description = ""
authors = ["Thinking Machines <hello@thinkingmachin.es>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.10.0"
fastapi = "^0.109.1"
uvicorn = "^0.21.1"
sse-starlette = "^1.3.3"
loguru = "^0.6.0"
openai = "^1.12.0"
pymysql = "^1.1.1"
google-search-results = "^2.4.2"
wolframalpha = "^5.0.0"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"
websockets = "^11.0.2"
pypdf = "^3.8.1"
tiktoken = "^0.7.0"
gptcache = "^0.1.22"
pycryptodome = "^3.18.0"
langchain = "^0.3.1"
azure-storage-blob = "^12.20.0"
azure-search-documents = "^11.6.0b4"
azure-identity = "^1.16.0"
cachetools = "^5.3.2"
asyncache = "^0.3.1"
msgraph-sdk = "^1.0.0"
azure-cognitiveservices-speech = "^1.37.0"
joserfc = "^0.9.0"
langchain-community = "^0.3.1"
langchain-openai = "^0.2.1"
pymssql = "^2.2.11"
pandas = "^1.4.4"
requests = "^2.32.0"
numpy = "^1.26.4"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.25.1"
alembic = "^1.13.2"
cuid2 = "^2.0.1"
cryptography = "^43.0.1"
pytest-cov = "^5.0.0"
deepeval = "^1.3.8"

[tool.poetry.dev-dependencies]
pre-commit = "~2.21.0"
pylint = "~2.15.10"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"
pytest = "^7.3.1"
black = "^24.3.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.black]
target-version = ["py310"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
pythonpath = "."
